simplify builder script
