+++ Lisensor.toml
@@ -2,7 +2,7 @@
 
 "packages/cli/**/*.rs" = "MIT"
 "packages/cli-build/**/*.rs" = "MIT"
-"packages/docker-build/examples/*.rs" = "MIT"
+"packages/docker-build/*.rs" = "MIT"
 
 "packages/lib/src/megaton/*.s" = "GPL-3.0-or-later"
 "packages/lib/src/megaton/*.c" = "GPL-3.0-or-later"
