--- packages/docker-build/builder/Taskfile.yml
@@ -1,35 +1,0 @@
-version: '3'
-
-includes:
-  docker:
-    taskfile: ../../../mono-dev/task/docker.yaml
-    internal: true
-    optional: true
-
-vars:
-  DOCKER_IMAGE: local.sh/megaton-docker/builder
-
-tasks:
-  build:
-    vars:
-      RUST_VERSION:
-        sh: rustc -V | sed -E 's/rustc ([^ ]+) .*$/\1/'
-    cmds:
-      - task: clean
-      - task: curl-rustup-init
-      - task: docker:build
-        dir: builder
-        vars:
-          ARGS: --build-arg RUST_VERSION={{.RUST_VERSION}}
-      - docker run --rm {{.DOCKER_IMAGE}} sh /root/builder-info.sh > temp/gcc-version.txt
-
-  clean:
-    - rm -rf temp/*
-    - mkdir -p temp
-
-  nuke:
-    - task: clean
-    - task: docker:nuke
-
-  curl-rustup-init:
-    - curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs > temp/rustup-init.sh
