+++ packages/toolchain-build/examples/docker_install.rs
@@ -1,5 +1,4 @@
 #[cu::cli]
 fn main(_: cu::cli::Flags) -> cu::Result<()> {
     megaton_toolchain_build::cmd::install(true /*keep*/, false /*clean*/)
-
 }
